Splits context render function from test suite
